requirements.txt
torch
transformers
spacy
accelerate
flask
openai
opencv-python
Pillow
pytesseract
pdf2image
python-docx